refactor(api-types): move workspace watcher WS event types to aionui-api-types
WatchChange / WatchChangeKind / WatchBatchEvent / WatchOverflowEvent are
WS event payloads, so per AGENTS.md they belong in aionui-api-types
alongside other request/response/WS DTOs, not inside aionui-file.

Pure relocation: definitions and serde derives unchanged, no behavior
change. aionui-file now imports them from aionui-api-types.
