+++ crates/aionui-api-types/src/lib.rs
@@ -25,6 +25,7 @@ mod system;
 mod team;
 mod team_mcp;
 mod websocket;
+mod workspace_watcher;
 
 pub use acp::{
     AcpEnvResponse, AcpHealthCheckRequest, AcpHealthCheckResponse, AgentModeResponse, DetectCliRequest,
@@ -133,3 +134,4 @@ pub use team::{
 };
 pub use team_mcp::{GuideMcpConfig, TeamMcpStdioConfig};
 pub use websocket::WebSocketMessage;
+pub use workspace_watcher::{WatchBatchEvent, WatchChange, WatchChangeKind, WatchOverflowEvent};
